pyproject.toml
[project]
name = "self-sensored"
version = "0.0.1"
description = "A REST API to receive Apple Healh data from the Auto Export iOS app."
authors = [{ name = "C. Thomas Brittain" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
